restructure proejct, add docker
