+++ Dockerfile
@@ -1,0 +1,59 @@
+# Multi-stage Dockerfile for Mimicking Mindsets
+FROM python:3.12-slim as base
+
+# Set environment variables
+ENV PYTHONUNBUFFERED=1 \
+    PYTHONDONTWRITEBYTECODE=1 \
+    PIP_NO_CACHE_DIR=1 \
+    PIP_DISABLE_PIP_VERSION_CHECK=1
+
+# Install system dependencies
+RUN apt-get update && apt-get install -y \
+    build-essential \
+    curl \
+    && rm -rf /var/lib/apt/lists/*
+
+# Create app user
+RUN groupadd -r appuser && useradd -r -g appuser appuser
+
+# Set work directory
+WORKDIR /app
+
+# Copy requirements first for better caching
+COPY requirements.txt .
+COPY pyproject.toml .
+
+# Install uv
+RUN pip install --no-cache-dir uv
+
+# Install Python dependencies using uv
+RUN uv pip install --system -r requirements.txt
+
+# Copy application code
+COPY . .
+
+# Create necessary directories
+RUN mkdir -p /app/logs /app/data
+
+# Change ownership to app user
+RUN chown -R appuser:appuser /app
+
+# Switch to non-root user
+USER appuser
+
+# Development stage
+FROM base as development
+ENV ENVIRONMENT=development
+EXPOSE 8000 5173
+CMD ["python", "web-interface/start_backend.py"]
+
+# Production stage
+FROM base as production
+ENV ENVIRONMENT=production
+EXPOSE 8000
+
+# Health check
+HEALTHCHECK --interval=30s --timeout=30s --start-period=5s --retries=3 \
+    CMD curl -f http://localhost:8000/health || exit 1
+
+CMD ["uvicorn", "web-interface.api_server:app", "--host", "0.0.0.0", "--port", "8000"]
