+++ pyproject.toml
@@ -1,7 +1,7 @@
 [project]
-name = "src"
+name = "mimicking-mindsets"
 version = "0.1.0"
-description = "Add your description here"
+description = "Multi-agent AI system simulating Turkish intellectuals Erol Güngör and Cemil Meriç"
 readme = "README.md"
 requires-python = ">=3.12"
 dependencies = [
@@ -22,4 +22,17 @@ dependencies = [
     "sentence-transformers>=4.1.0",
     "tabulate>=0.9.0",
     "uvicorn>=0.34.3",
+]
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.uv]
+dev-dependencies = [
+    "pytest>=7.0.0",
+    "pytest-asyncio>=0.21.0",
+    "black>=23.0.0",
+    "isort>=5.12.0",
+    "flake8>=6.0.0",
 ]
