+++ tests/__init__.py
@@ -1,0 +1,7 @@
+"""
+Test suite for Mimicking Mindsets multi-agent system.
+
+This package contains tests for:
+- Individual persona agents (Phase 1)
+- Multi-agent orchestration (Phase 2)
+"""
