+++ web-interface/frontend/src/components/AgentTraces.jsx
